Ability to play the game added
Created the ability to click and place a nought or cross depending on
player, the game now updates on click to show where this was placed and
check if there was a winner or draw. If there is no winner or draw, game
continues. Otherwise, GameState will update appropriately to paint
correct message on screen and restart game on click.